+++ XuanYu.Editor.UI/Top/Top.axaml.cs
@@ -1,4 +1,5 @@
 using Avalonia.Controls;
+using Avalonia.Interactivity;
 
 namespace XuanYu.Editor.UI;
 
@@ -7,5 +8,10 @@ public partial class Top : UserControl
     public Top()
     {
         InitializeComponent();
+    }
+
+    async void RegionDrawing_Click(object? sender, RoutedEventArgs e)
+    {
+        if (DataContext is UiVm vm) await vm.BeginRegionDrawingAsync();
     }
 }