+++ XuanYu.Editor.UI/Vm/Mode/UiVm.Mode.cs
@@ -37,6 +37,7 @@ public sealed partial class UiVm
         OnPropertyChanged(nameof(IsEditMode)); OnPropertyChanged(nameof(IsMapEditMode));
         OnPropertyChanged(nameof(IsRegionEditMode)); OnPropertyChanged(nameof(CurrentEditorModeText));
         OnPropertyChanged(nameof(CanStartRegionDrawing));
+        OnPropertyChanged(nameof(CanRequestRegionDrawing));
         OnPropertyChanged(nameof(IsMapEditorMode));
         RaiseLayerContextBindings();
     }